move extractActivePorts into datastore
