+++ pkg/epp/util/pod/pod_test.go
@@ -17,134 +17,12 @@ limitations under the License.
 package pod
 
 import (
-	"reflect"
 	"testing"
 	"time"
 
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/util/sets"
 )
-
-func TestExtractActivePorts(t *testing.T) {
-	tests := []struct {
-		name           string
-		pod            *corev1.Pod
-		expectedPorts  sets.Set[int]
-		expectedExists bool
-	}{
-		{
-			name: "Pod without active ports annotation",
-			pod: &corev1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:        "test-pod",
-					Namespace:   "default",
-					Annotations: map[string]string{},
-				},
-			},
-			expectedPorts:  nil,
-			expectedExists: false,
-		},
-		{
-			name: "Pod with empty active ports annotation",
-			pod: &corev1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:        "test-pod",
-					Namespace:   "default",
-					Annotations: map[string]string{activePortsAnnotation: ""},
-				},
-			},
-			expectedPorts:  sets.New[int](),
-			expectedExists: true,
-		},
-		{
-			name: "Pod with single port in annotation",
-			pod: &corev1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:        "test-pod",
-					Namespace:   "default",
-					Annotations: map[string]string{activePortsAnnotation: "8000"},
-				},
-			},
-			expectedPorts:  sets.New[int](8000),
-			expectedExists: true,
-		},
-		{
-			name: "Pod with multiple ports in annotation",
-			pod: &corev1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:        "test-pod",
-					Namespace:   "default",
-					Annotations: map[string]string{activePortsAnnotation: "8000,8001,8002"},
-				},
-			},
-			expectedPorts:  sets.New[int](8000, 8001, 8002),
-			expectedExists: true,
-		},
-		{
-			name: "Pod with multiple ports with spaces in annotation",
-			pod: &corev1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:        "test-pod",
-					Namespace:   "default",
-					Annotations: map[string]string{activePortsAnnotation: "8000, 8001 , 8002"},
-				},
-			},
-			expectedPorts:  sets.New[int](8000, 8001, 8002),
-			expectedExists: true,
-		},
-		{
-			name: "Pod with invalid port in annotation (non-numeric)",
-			pod: &corev1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:        "test-pod",
-					Namespace:   "default",
-					Annotations: map[string]string{activePortsAnnotation: "8000,invalid,8002"},
-				},
-			},
-			expectedPorts:  sets.New[int](8000, 8002),
-			expectedExists: true,
-		},
-		{
-			name: "Pod with invalid port in annotation (negative number)",
-			pod: &corev1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:        "test-pod",
-					Namespace:   "default",
-					Annotations: map[string]string{activePortsAnnotation: "8000,-1,8002"},
-				},
-			},
-			expectedPorts:  sets.New[int](8000, 8002),
-			expectedExists: true,
-		},
-		{
-			name: "Pod with duplicate ports in annotation",
-			pod: &corev1.Pod{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:        "test-pod",
-					Namespace:   "default",
-					Annotations: map[string]string{activePortsAnnotation: "8000,8001,8000"},
-				},
-			},
-			expectedPorts:  sets.New[int](8000, 8001),
-			expectedExists: true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			ports, exists := ExtractActivePorts(tt.pod)
-
-			if exists != tt.expectedExists {
-				t.Errorf("ExtractActivePorts() exists = %v, want %v", exists, tt.expectedExists)
-			}
-
-			if !reflect.DeepEqual(ports, tt.expectedPorts) {
-				t.Errorf("ExtractActivePorts() ports = %v, want %v", ports, tt.expectedPorts)
-			}
-		})
-	}
-}
 
 func TestIsPodReady(t *testing.T) {
 	tests := []struct {
